+++ libsidplay2/sidtune/SidTuneTools.cpp
@@ -87,7 +87,7 @@ char* SidTuneTools::slashedFileNameWithoutPath(char* s)
 // The backwards-version.
 char* SidTuneTools::fileExtOfPath(char* s)
 {
-    uint_least32_t last_dot_pos = strlen(s);  // assume no dot and append
+    uint_least32_t last_dot_pos = (uint_least32_t)strlen(s);  // assume no dot and append
     for ( int pos = last_dot_pos; pos >= 0; --pos )
     {
         if ( s[pos] == '.' )
